third_party/WebKit/LayoutTests/fast/forms/number/number-type-update-by-change-event.html

<!DOCTYPE>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<input type="number" onchange="handleChange(this);">
<script>
function sendKey(keyName) {
    eventSender.keyDown(keyName);
}

function handleChange(element) {
    element.type = '';
    ++changeEventCounter;
}

var changeEventCounter = 0;
var numberInput = document.getElementsByTagName('input')[0];
numberInput.focus();
sendKey('ArrowUp');

testPassed('if not crashed in ASAN build.');
shouldBe('changeEventCounter', '1');
</script>
</body>
</html>
